Moving ActivityProvider to device module
