+++ features/in-app-update/src/main/java/br/com/recipebook/inappupdate/view/InAppUpdaterImpl.kt
@@ -1,10 +1,10 @@
 package br.com.recipebook.inappupdate.view
 
 import androidx.fragment.app.FragmentActivity
+import br.com.recipebook.device.ActivityProvider
 import br.com.recipebook.inappupdate.InAppUpdateResult
 import br.com.recipebook.inappupdate.domain.InAppUpdater
 import br.com.recipebook.inappupdate.requestInAppUpdate
-import br.com.recipebook.view.ActivityProvider
 import com.google.android.play.core.install.model.AppUpdateType
 
 private const val REQUEST_CODE = 1292
