+++ infrastructure/configuration/src/main/java/br/com/recipebook/di/ConfigurationModule.kt
@@ -7,7 +7,6 @@ import br.com.recipebook.data.remote.ConfigurationApi
 import br.com.recipebook.data.remote.ConfigurationDataSourceRemote
 import br.com.recipebook.data.remote.ConfigurationDataSourceRemoteImpl
 import br.com.recipebook.domain.ConfigurationRepository
-import br.com.recipebook.view.ActivityProvider
 import org.koin.dsl.module
 import retrofit2.Retrofit
 
@@ -30,11 +29,6 @@ val configurationDataModule = module {
     }
 }
 
-val configurationViewModule = module {
-    single { ActivityProvider(get()) }
-}
-
 val configurationModules = listOf(
-    configurationDataModule,
-    configurationViewModule
+    configurationDataModule
 )
