+++ infrastructure/device/src/main/java/br/com/recipebook/device/LocaleProvider.kt
@@ -1,0 +1,15 @@
+package br.com.recipebook.device
+
+import java.util.Locale
+
+private val SUPPORTED_LANGUAGES = listOf("en", "pt")
+
+interface LocaleProvider {
+    fun getLocale(): String
+}
+
+internal class LocaleProviderImpl : LocaleProvider {
+    override fun getLocale(): String {
+        return Locale.getDefault().language.takeIf { SUPPORTED_LANGUAGES.contains(it) } ?: "en"
+    }
+}
